initial commit
This is an attempt at an Ansible role, designed to let you manage a
number of guests and networks on a KVM host, as defined by YAML
inventories.

It supports custom guest settings, creating the disk images and setting
the guest to a valid state, such as running, shutdown, destroyed or
undefined. It can also create NAT networks and put the guests on any of
those networks.

It uses cloud images from major distros and has been tested with Fedora,
CentOS, Debian, Ubuntu and openSUSE.

A working KVM host is required, again tested on the above distros,
however most distros with a libvirt and KVM probably work.

There is an example Ansible play which uses this role here which might
help:

https://github.com/csmart/virt-infra-ansible

See README.md for more details.