Register RHEL guests to ports for package install
If you're using RHEL guest, installing packages (including cloud-init)
requires the guest to be registered to the Red Hat portal.

To support this properly, the virt-customise needs to have subscription
manager variables (See
http://libguestfs.org/virt-builder.1.html#subscription-manager).

Thus, if the guest is RHEL you must include the new
`virt_infra_sm_creds` variable.

Here is an example using a password:
  virt_infra_sm_creds: "portal-username:password:portal-password"

Here is an example using a password file:
  virt_infra_sm_creds: "portal-username:file:/path/to/file"

By default this will auto attach, however if you want to specify a
specific pool then set it with the `virt_infra_sm_attach` variable.

The node will be cleaned and removed from the portal post package
install.