add disks in a clearer way
There was an issue previously where cloud-init did not discover the ISO
on boot, causing it to fail.

To fix this, the CDROM device was added directly after root.

However in subsequent tests with lots of disks, it appears extra
controllers are added and the CDROM goes onto a different LSI one. This
appears to be behind the CDROM not being detected by cloud-init.

Thus, to fix this, we're adding two controllers both of type
virtio-scsi. This way both the disks and CD-ROM are on the same
controller type and it improves detection at boot.

This also means that adding the boot disk and CDROM don't need to be
treated specially and boot can be done in the loop.

This also fixes an issue where boot is not the first disk listed in the
vm's disks, by always setting it to be the first boot device.

If you are adding more than 13 disks however, we will need to add
another controller. I'm not sure anyone will hit this, if they do I'll
address it then.