support ssd for scsi disks and dynamic controllers
SCSI disks can set new variable of 'ssd = true' in the YAML inventory in
order to set them as non-rotational disks (SSDs).

In order to enable SSD support we must know the alias of the disk, which
unfortunately we are not able to define ourselves with virt-install.

The alias is generated from the controller number and the disk location
on that controller. For example, 'scsi0-0-0-1' is the second device on
the first controller. This format is only true for the virtio-scsi
controller, the LSI controller would have a slightly different format
'scsi0-0-1' for the same disk.

In order to calculate the alias, we need to know which controller the
disk is on and its location. And in order to do that, we need to count
the number of SCSI disks as well as the controllers.

The SCSI controllers have a max of 7 devices before a second controller
is required. Thus, we need to add a second controller once the number of
SCSI disks reaches 7.

Lastly, even though a controller supports up to 7 disks, if there is
more than 3, then libvirt puts the CD-ROM onto a new controller.
Unfortunately this is LSI by default which means your disk and the
CD-ROM are not on the same bus. This can result in detection or timing
issues, causing cloud-init to fail as the CD-ROM is not detected before
the system keeps booting from the OS disk.

Therefore, we have to add an extra virtio-scsi controller when the
number of disks reaches more than 3, so that the CD-ROM is on the same
bus as the root disk.

To make this work I had to be able to increment a counter in Ansible.
The only way I could do that reliably on anything less than Jinja 2.10
(which solves scope issue) is to use a list and its length. It might not
be pretty, but it works. I'll gladly buy a few beverages for anyone who
shows me a better way to do it!