add support for libvirt OVS networks
Thanks to my friend Nick Satsia (nsatsia) for his help with this!

This add support for creating OVS networks in libvirt and attaching VMs
to them. Note that the OVS bridge needs to already exist on the kvmhost
(see below).

This configuration is flexible, so how you configure it is up to you.
However, you might like to have one libvirt OVS bridge with multiple
portgroups (which are what guests connect to).

One portgroup could be a trunk where you include all if your VLANs,
perhaps with a native. This way you can create tagged interfaces in
Linux.

Then you can also create other portgroups which are specific to VLANs
and networks. This way you can attach the VM to multiple separate
networks directly without needing tagged interfaces in Linux.

If you don't use VLANs you can leave them off altogether and just use it
as a bridge to your native network.

Define the libvirt OVS bridge to create in the kvmhost YAML inventory
file. The bridge_dev is important, this is the ovs bridge to connect to.
As above, here's an example with a trunk and tagged networks.

  virt_infra_host_networks:
    present:
     - name: ovs-bridge
       type: ovs
       bridge_dev: ovs-bridge
       portgroup:
         - name: ovs-trunk
           trunk: true
           vlan:
             - 1
             - 2
         - name: ovs-vlan1
           native_vlan: 1
           vlan:
             - 1
         - name: ovs-vlan2
           native_vlan: 2
           vlan:
             - 2

Then in the YAML inventory for the VM, define the OVS portgroups you
want to the VM to connect to. Remember the first interface probably
should be something with a native VLAN so that DHCP works.

  test-vm:
    virt_infra_networks:
      - name: ovs-bridge
        portgroup: ovs-vlan1
        type: ovs
      - name: ovs-bridge
        portgroup: ovs-vlan2
        type: ovs

Note this requires openvswitch to be installed and running with a OVS
bridge *already* created. A physical interface should also already be
added if it needs to communicate to the rest of the network.

You can configure this easily with ovs-vsctl, but it doesn't survive a
reboot:

 $ sudo ovs-vsctl add-br ovs-bridge
 $ sudo ovs-vsctl add-port ovs-bridge eth0

Or configure persistently with NetworkManager:

 $ sudo nmcli con add type ovs-bridge conn.interface ovs-bridge con-name ovs-bridge
 $ sudo nmcli con add type ovs-port conn.interface port-ovs-bridge master ovs-bridge
 $ sudo nmcli con add type ovs-interface slave-type ovs-port conn.interface ovs-bridge master port-ovs-bridge

 $ sudo nmcli con modify id eth0 ipv4.method disabled ipv6.method disabled
 $ sudo nmcli con add type ovs-port conn.interface ovs-port-eth master ovs-bridge con-name ovs-port-eth
 $ sudo nmcli con add type ethernet conn.interface eth0 master ovs-port-eth con-name ovs-int-eth
 $ sudo ifup ovs-int-eth

Or some other way using your distro's configs.