add support for custom VM MAC addresses
It may be useful to specify a MAC address for VMs so that they are
always the same when recreated. For example, when they are used for
external DHCP leases on a bridged network, or for PXE boot, etc.

This adds a new "mac" YAML option for VM network interfaces and works
with every network type, e.g.:

  test-0:
    virt_infra_networks:
      - name: br0
        type: bridge
        mac: 52:54:00:aa:bb:cc