fix scsi count
The default bus is scsi, so if it is not specified we still need to
increment the scsi counter. This was causing an extra controller to be
added when it wasn't needed.

This patch increments the counter when bus is scsi or not specified, and
fixes the number of disks required before a new controller is required
for the CD-ROM drive.