add support for keeping disks
This change adds the ability to keep a disk rather than delete it when a
guest is undefined.

This adds a new 'keep' disk config item. Setting this to true will cause
the disk to be skipped when the machine is undefined.

If this is the boot disk, then if you re-create the same guest after
it's deletion, the disk will not have sysprep run on it, or cloud-init,
etc installed.

Thus, you can set 'keep = true' on the boot disk and after the first
time it's created it will remain the same over subsequent guest deletes
and recreates.