add support for virtual bmc
This enables option for creating and removing vbmc interfaces for
guests, so that they can be controlled with IPMI.

By default vbmc is disabled, set this in kvmhost vars to enable it:

  virt_infra_vbmc: true

If enabled, by default this will install the latest virtualbmc from pip and
configure it to run as root. You may specify a specific version of
virtualbmc, e.g. 1.6.0, with the following var:

  virt_infra_vbmc_pip_version: 1.6.0

However, Fedora and openSUSE do package virtualbmc, so if you'd rather
use the system package (or install it manually), you can disable pip
install:

  virt_infra_vbmc_pip: false

The default service is called "vbmcd" so if it is called something else
on your distro (like "virtualbmc") then specify it with the following on
kvmhost:

  virt_infra_vbmc_service: virtualbmc

The Ansible will do some simple checks to make sure that the port is not
already in use by a different host.

This adds some new vars for guests, the only required one is `vbmc_port`
as each VM needs to be on a dedicated UDP port on the KVM host.

  ipmi-guest:
    virt_infra_vbmc_port: 62300
    virt_infra_vbmc_user: admin
    virt_infra_vbmc_password: password

You can see your running virtual BMC interfaces with vbmc tool.

 $ sudo vbmc list
 +-------------+---------+---------+-------+
 | Domain name | Status  | Address |  Port |
 +-------------+---------+---------+-------+
 | ipmi-guest  | running | ::      | 62300 |
 +-------------+---------+---------+-------+

You should now be able to interact with it using ipmitool.

 $ ipmitool -I lanplus -H localhost -p 62300 -U admin -P password power status
 Chassis Power is on

Finally, this doesn't automatically punch a hole in your firewall, so if
you have one you might want to open the range you use.

 $ sudo firewall-cmd --zone libvirt --add-port=62300-62399/udp