add support for qemu security driver, fix NVMe on apparmor
Most distros implement mandatory access controls like SELinux and
Apparmor around libvirt. However, NVMe drives are attached via extra
qemu commands and this causes launching of guests to fail as access to
the drives is blocked.

This change adds a new variable `virt_infra_security_driver` which lets
you set the mandatory access control driver you want to use. By default
this is set to "none" which disables it, however on RPM distros (RHEL,
CentOS, Fedora, openSUSE) this is overridden to "selinux" and on Deb
based distros (Debian, Ubuntu) to "apparmor" so it is still enabled by
default.

On Fedora, setting the right ownership of the drives is enough to make
this work with SELinux.

However, this is not enough on Debian and Ubuntu where Apparmor still
blocks reading with an error like this:

  audit: type=1400 audit(1588465721.729:1232):
    apparmor="DENIED" operation="open"
    profile="libvirt-c814c482-e1f0-4c47-ba9f-2b354f89d0e9"
    name="/var/lib/libvirt/images/example-ubuntu-eoan-data-nvme.qcow2"
    pid=1700 comm="qemu-system-x86" requested_mask="r" denied_mask="r"
    fsuid=64055 ouid=64055

Therefore, when virt_infra_security_driver is set to "apparmor" we add a
new rule to /etc/apparmor.d/abstractions/libvirt-qemu which allows VMs
to access to NVMe drives. It's probaby better to have custom profile for
the VM created with virt-aa-helper, however I'm not an Apparmor expert
and this is just designed for dev work (patches welcome!).

For details on Apparmor integration with libvirt, see doco:
https://gitlab.com/apparmor/apparmor/-/wikis/Libvirt