fix boot order conflict on Debian
Originally we told virt-install to boot from hard drive (--boot hd)
however as a part of commit bc99a15f[1], "add disks in a clearer way, we
now specify a boot order when defining the disks.

On Debian, having both caused a conflict and the guests could not be
defined. This change removes the --boot option as we can now just depend
on the disk boot order.

[1] https://github.com/csmart/ansible-role-virt-infra/commit/bc99a15f095d78c916506dd7e3fe1d0d6db48226