ensure netfilter disabled on bridges
When VMs are connected to a bridge they can be sent to iptables for processing.

This is generally not desired for VM on a bridge, as it will block
traffic coming in to the VM.

This patch makes sure that, if br_netfilter exists, it is loaded and
then disabled on bridges with sysctl.