pass kvmhost check when --limit set to "all"
The kvmhost must be included in the limit so that the guests can be
created. A check exists to make sure this is included, however if limit
is set to "all" it will also include kvmhost, so we should not fail in
this case (it's the same as excluding a limit).