add MTU support on libvirt and guest networks
This patch allows you to set the MTU for libvirt networks on KVM host as
well as for guest interfaces.

Simply add the `mtu` option to your kvmhost network definitions and/or
your guest network interfaces definitions. When setting on a libvirt
host network and specifying an IP, that interface will get that MTU.

Also note that setting the MTU for an OVS host network is not supproted
by libvirt and will be ignored. However, setting an MTU for a guest
connected to an OVS network is supported.

Note that if you're connecting your guest to a bridge, the bridge on the
KVM host should also have been set to the same or higher MTU.