turn guest deps into variable
By default we make sure that cloud-init and qemu-guest-agent are
installed into the guest disk image so that the VM can be configured on
boot. However, if the image already includes these packages, then this
step can be skipped.

Furthermore, you might want to have different or additional packages
installed.

This patch introduces a list variable of `virt_infra_guest_deps` which
defaults to the existing values:

  virt_infra_guest_deps:
    - cloud-init
    - qemu-guest-agent

To skip the step which installs these packages, set this to an empty
list in your inventory, e.g.

  virt_infra_guest_deps: []

If you want to install different packages then make a new list, e.g.:
  virt_infra_guest_deps:
    - cloud-init
    - git
    - qemu-guest-agent
    - rsync
    - vim

As with all vars, these can be at the host or group level.

This patch also includes vars/rhel.yml which was missing for RHEL hosts.