add support for cloned boot disk instead of backed
By default the boot disk is created using the specified distro disk
image as a backing file.

This commits add a new boolean disk option, `clone` which will copy the
source distro disk to be used as the boot disk image instead.

Due to the way that convert and create subcommands from qemu-img work,
the size of the boot disk is now done as a separate step.

Additional disks are also now done as a separate step.

This commit should also pave the way for using RAW disk images instead
of qcow2, as they do not support backing files.