add support for network config via cloud-init
This patchset enables support for basic network configuration v2 using
cloud-init. Currently only ethernet devices are supported (not bonds,
bridges or VLANs).

For details, see the cloud-init documentation:
https://cloudinit.readthedocs.io/en/latest/topics/network-config-format-v2.html

As not all distros may support this, in order to activate network config
please enable the following variable:

  virt_infra_network_config: true

The network config is then populated with new optional network
configuration options for each VM's network interfaces, including:

 - device (string, to set the interface device name, e.g. eth0)
 - addresses (list)
 - dhcp4 (boolean, disabled if addresses is defined, otherwise true)
 - dhcp6 (boolean, disabled if addresses is defined, otherwise true)
 - gateway4 (string, IPv4 default gateway)
 - gateway6 (string, IPv6 default gateway)
 - nameservers (dict)
   - search (list of DNS domain search)
   - addresses (list of DNS servers)
 - routes (list of dicts)
   - to (key in dict, value is subnet to route to)
   - via (key in dict, value is gateway address/subnet)
   - metric (key in dict, int)

NOTE: Due to the way cloud-init works (see documentation link above), if
DHCP is enabled then some manual settings (including gateway and
nameservers) are ignored. Manual routes are kept.

It is worth noting that none of the new options are actually required
(defaults to eth0 and dhcp), but here is an example:

  virt_infra_networks:
    - name: default
      device: eth0
      addresses:
        - 192.168.0.123/24
      gateway4: 192.168.0.1
      nameservers:
        search: [foo.local, bar.local]
        addresses: [8.8.8.8]
      routes:
        - to: 192.0.2.0/24
          via: 11.0.0.1
          metric: 3

The network config format requires specifying a network device to
configure which matches the actual interface. Unfortunately, there's no
way to always know that, as some distros will enable persistent names
and others won't.

For example, if we say "eth0" but the interfaces are actually "ens0"
instead, then the config will still be written to eth0 and therefore not
apply to ens0.

The solution to this is to rename the interface, so that it will match,
however in order to ensre we pick the exact interface we must match on
MAC address.

So, if a MAC address is not specified for the interface, this change
also adds a way to ensure that VMs can use network config reliably by
deterministically generating MAC addresses. As an ugly hack, it does
this based on the interface's place in the list of networks
(loop.index0), plus the standard random_mac method based on inventory
hostname, e.g.

  '{{ network.mac | default(('52:54:' + ("%02d" | format(loop.index0) |string)) | random_mac(seed=inventory_hostname)) }}'

If you have a better way to do this, I gladly accept patches!

Finally, the config must have a valid entry, it cannot be a blank file
and it cannot just have boilerplate like this:

  version: 2
  ethernets:

Furthermore, if we have a MAC address specified then it must match a
real interface on the box. Hence we generate them in the config if not
specified and ensure the VM gets the same addresses on the same
interfaces.

If either of these requirements are not met, cloud-init crashes and
fails to read any other config so the VM fails to boot successfully.