always resize boot drive
Recently I added the ability to use cloned boot drives and at the time I
thought it was best to keep the same size as the source, unless
overridden with `size` variable.

However, this approach meant that while you could rely on the default or
overridden value of `virt_infra_disk_size` for all subsequent drives,
you would have to also explicitly set `size` for the cloned boot disk if
you wanted it to be the same as the rest.

Many cloud source images are deliberately small and expected to be
resized, thus you could end up with a boot disk smaller than expected.

This change flips the logic, so that boot is always resized to the
default, via `virt_infra_disk_size` variable or `size` variable in the
disk definition, if explicitly set. This brings it inline with the rest
of the disks.

So now, if you want your boot disk to be the same size as the source,
set it explicitly with disk's `size` variable instead.