fix ssh config when keys are specified
When the code for writing ~/.ssh/config was updated to support multiple
KVM hosts, a mistake was made which also wrote SSH keys defined in the
inventory. However, SSH config expects a path to a file, not the actual
key.

This breaks the ability to SSH into the VMs as the config is not valid.

This patch reverts to previous code, so that it once again only writes
the identity files in SSH config if none are specified (and thus use
the path to SSH keys which were found on the host).

Thanks to @benformosa for reporting the issue.