add support for custom host packages
Sometimes your KVM hosts might require some additional packages, so this
introduces a new list variable `virt_infra_host_pkgs_custom` which will
install packages in the list.

Note that the variable must be a list and the host must be able to
install the packages, i.e. the package names must be correct for the
host and if a RHEL host, it must be registered.